Move sampling_ratio tests to tests/tracing/test_fluent.py
Move manual tests from test_sampling_ratio.py to the proper test location
and convert them to pytest format following MLflow test conventions.
